+++ LocalAuthenticationWrapper/Sources/LocalAuthenticationWrapper/LocalAuthWrap.swift
@@ -1,0 +1,7 @@
+public protocol LocalAuthWrap {
+    associatedtype LAType: LocalAuthType
+    var type: LAType { get }
+    
+    func checkMinimumLevel(_ requiredLevel: any LocalAuthType) -> Bool
+    func enrolLocalAuth() async throws -> Bool
+}
